moved parseResponse to server utils folder
